+++ slashing-protection/src/main/java/tech/pegasys/web3signer/slashingprotection/validator/AttestationValidator.java
@@ -64,9 +64,9 @@ public class AttestationValidator {
           sourceEpoch,
           targetEpoch,
           publicKey);
-      return false;
+      return true;
     }
-    return true;
+    return false;
   }
 
   public boolean existsInDatabase() {