+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 - Interlock/Armory II HSM keystore support
+- Eth2 slashing protection data able to be imported from json file (Interchange format V5)
 
 ## 0.2.0